Reimplemented responder framework - NOT using Avro/IPC Responders
Avro RPC API is not particularly designed for full reuse when the wire
format is not compliant with Avro framing. For example, in order to
support Content-Type "application/json", "application/x-protobuf" or
"application/x-thrift", one has to write ones own Responder and cannot
rely on the Avro implementation. This is because the Avro specific
implementation assumes that the "on the wire" framing and handshake is
Avro style. Clearly that is only useful if Content-Type was
"avro/binary". The new DemandSideServer implementation provides a much
simpler/ cleaner reuse of Avro / IO SDK without creating dependency on
Avro/ IPC.