Rearchitected the DSP Configuration framework
DemandSideServer must read in from configuration file (or Database)
which exchanges are authorized to send BidRequest feed (a.k.a. supply
feed), what Content-Type will be in use and other SSP specific
parameters. It must also read in the list of Advertisers using this DSP
as a bidding front end, which advertisers have Seats with which
exchanges, and any other Advertiser specific information. All this has
been rearchitected using the DAO pattern to remove hard dependency on
config or property files. Initial implementation of the DemandSideDAO
interface is for a JsonFileBackedDAO.